Extract workspace configuration integration tests
